pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-learning-materials"
version = "0.0.0"
description = "SQL scripts that demonstrate various features and concepts."
authors = [{name = "Bilbottom"}]
readme = "README.md"
keywords = ["sql", "sql-server", "postgres", "sqlite", "duckdb"]
urls = {Repository = "https://github.com/Bilbottom/sql-learning-materials"}
requires-python = ">=3.11,<4.0"
dependencies = [
    "db-query-profiler>=0.0.6",
    "python-dotenv>=1.1.0",

    # DuckDB driver
    "duckdb>=1.1.3",
    # PostgreSQL driver
    "psycopg2>=2.9.9; sys_platform == 'win32'",
    "psycopg2-binary>=2.9.9; sys_platform == 'darwin'",
    # SQL Server driver
    "pyodbc>=5.1.0",
    # Snowflake driver
    "snowflake-connector-python>=3.14.0",
]

[dependency-groups]
dev = [
    "coverage-badge>=1.1.0",
    "pre-commit>=4.1.0",
]
test = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]
docs = [
    "markdown-callouts>=0.4.0",
    "mdx-truly-sane-lists>=1.3",
    "mkdocs>=1.6.1",
    "mkdocs-callouts>=1.16.0",
    "mkdocs-material>=9.6.7",
]


[tool.setuptools]
packages = ["src"]


[tool.pytest.ini_options]
addopts = "--cov=src --cov-report term-missing"
testpaths = ["tests"]


[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["F", "I", "N", "PL", "R", "RUF", "S", "UP", "W"]
ignore = []
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# https://github.com/astral-sh/ruff/issues/4368
[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    "S101",    #  Use of `assert` detected
    "PLR2004", #  Magic value used in comparison
    "PLR0913", #  Too many arguments in function definition
]
